SUMMARY:   fixed typecast bug in managememory.
     BUGS:      hopefully fixes Aztecoo Bug 3754 (unsigned integer cast)

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  No

     3) Did you rerun the test suite on at least the packages that you
        changed?  Not really. I ran one problem and it worked. Overall, this should only effect very large problems (which we probably don't have in the testsuite).

     COMMENTS:

     Unsigned int was being cast to an int as opposed to a long int.
     Probably this whole routine should be changed ... but this should help. Basically, the unsigned int variable was meant to handle slightly larger numbers (but this was lost when casting it to an int).

(checklist version 2007-06-20)

+++ packages/aztecoo/src/az_util.c
@@ -929,7 +929,7 @@ int aligned_str_mem,aligned_j,aligned_size;
 double *dtmp;
 
   /**************************** execution begins ******************************/
-  size = (int) input_size;
+  size = (long int) input_size;
   current = head;
 
   if (action == -43) {
